+++ app/Http/Controllers/Api/Project/RevenuesKwhController.php
@@ -405,7 +405,7 @@ class RevenuesKwhController extends ApiController
         return FullRevenueDistributionKwh::collection($distribution);
     }
 
-    public function downloadPreview(Request $request, RevenueDistributionKwh $distributionKwh)
+    public function previewPDF(Request $request, RevenueDistributionKwh $distributionKwh)
     {
         //get current logged in user
         $user = Auth::user();
@@ -474,6 +474,7 @@ class RevenuesKwhController extends ApiController
 
             $revenueHtml
                 = TemplateVariableHelper::stripRemainingVariableTags($revenueHtml);
+
             $pdf = PDF::loadView('documents.generic', [
                 'html' => $revenueHtml,
             ])->output();
@@ -532,7 +533,6 @@ class RevenuesKwhController extends ApiController
 
                 $email = Mail::fromMailbox($mailbox)
                     ->to($primaryEmailAddress->email);
-
                 if (!$subject) {
                     $subject = 'Participant rapportage Econobis';
                 }
@@ -609,7 +609,7 @@ class RevenuesKwhController extends ApiController
         set_time_limit(0);
         $distributionKwhIds = $request->input('distributionKwhIds');
         $subject = $request->input('subject');
-        $documentTemplateId = $request->input('documentTemplateId');
+        $documentTemplateId = ( $request->input('documentTemplateId') && !empty($request->input('documentTemplateId')) ) ?? null;
         $emailTemplateId = $request->input('emailTemplateId');
         $showOnPortal = $request->input('showOnPortal');
 
@@ -650,7 +650,7 @@ class RevenuesKwhController extends ApiController
         return null;
     }
 
-    public function createParticipantRevenueReport($subject, $distributionKwhId, DocumentTemplate $documentTemplate, EmailTemplate $emailTemplate, $showOnPortal)
+    public function createParticipantRevenueReport($subject, $distributionKwhId, ?DocumentTemplate $documentTemplate, EmailTemplate $emailTemplate, $showOnPortal)
     {
         $portalName = PortalSettings::get('portalName');
         $cooperativeName = PortalSettings::get('cooperativeName');
@@ -662,133 +662,131 @@ class RevenuesKwhController extends ApiController
 
         $messages = [];
 
-        //load template parts
-        $documentTemplate->load('footer', 'baseTemplate', 'header');
-
-        $html = $documentTemplate->header ? $documentTemplate->header->html_body : '';
-
-        if ($documentTemplate->baseTemplate) {
-            $html .= TemplateVariableHelper::replaceTemplateTagVariable($documentTemplate->baseTemplate->html_body,
-                $documentTemplate->html_body, '', '');
-        } else {
-            $html .= TemplateVariableHelper::replaceTemplateFreeTextVariables($documentTemplate->html_body,
-                '', '');
-        }
-
-        $html .= $documentTemplate->footer
-            ? $documentTemplate->footer->html_body : '';
-
         $distributionKwh = RevenueDistributionKwh::find($distributionKwhId);
+        $contact = $distributionKwh->contact;
+        $revenuesKwh = $distributionKwh->revenuesKwh;
+        $project = $revenuesKwh->project;
+        $administration = $project->administration;
+        $orderController = new OrderController();
+        $contactInfo = $orderController->getContactInfoForOrder($contact);
+        $primaryEmailAddress = $contact->primaryEmailAddress;
+
+        $document = null;
 
         if( !( empty($distributionKwh->address)
             || empty($distributionKwh->postal_code)
             || empty($distributionKwh->city) ) ) {
 
-            $contact = $distributionKwh->contact;
-            $orderController = new OrderController();
+            if($documentTemplate) {
+                //load template parts
+                $documentTemplate->load('footer', 'baseTemplate', 'header');
 
-            $contactInfo = $orderController->getContactInfoForOrder($contact);
-            $subject = str_replace('{contactpersoon}', $contactInfo['contactPerson'], $subject);
+                $html = $documentTemplate->header ? $documentTemplate->header->html_body : '';
 
-            $primaryEmailAddress = $contact->primaryEmailAddress;
-
-            $revenuesKwh = $distributionKwh->revenuesKwh;
-            $project = $revenuesKwh->project;
-            $administration = $project->administration;
-
-            $html = str_replace('{contactpersoon}', $contactInfo['contactPerson'], $html);
-
-            $revenueHtml = TemplateTableHelper::replaceTemplateTables($html, $contact);
-
-            $revenueHtml
-                = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'contact', $contact);
-            $revenueHtml = TemplateVariableHelper::replaceTemplatePortalVariables($revenueHtml,'portal' );
-            $revenueHtml = TemplateVariableHelper::replaceTemplatePortalVariables($revenueHtml,'contacten_portal' );
-            $revenueHtml = TemplateVariableHelper::replaceTemplateCooperativeVariables($revenueHtml,'cooperatie' );
-
-            //wettelijk vertegenwoordiger
-            if (OccupationContact::where('contact_id', $contact->id)->where('occupation_id', 7)->exists()) {
-                $wettelijkVertegenwoordiger = OccupationContact::where('contact_id', $contact->id)
-                    ->where('occupation_id', 7)->first()->primaryContact;
-                $revenueHtml
-                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'wettelijk_vertegenwoordiger', $wettelijkVertegenwoordiger);
-            }
-            $revenueHtml
-                = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'ik', $user);
-            $revenueHtml
-                = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'administratie', $administration);
-            $revenueHtml
-                = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'verdeling', $distributionKwh);
-            $revenueHtml
-                = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'opbrengst', $revenuesKwh);
-            $revenueHtml
-                = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'project', $project);
-            $revenueHtml
-                = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'deelname', $distributionKwh->participation);
-            $revenueHtml
-                = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'mutaties', $distributionKwh->participation->mutations);
-
-            $revenueHtml
-                = TemplateVariableHelper::stripRemainingVariableTags($revenueHtml);
-            $pdf = PDF::loadView('documents.generic', [
-                'html' => $revenueHtml,
-            ])->output();
-
-            try
-            {
-                $time = Carbon::now();
-
-                $documentCreatedFromParticipantId = DocumentCreatedFrom::where('code_ref', 'participant')->first()->id;
-
-                $document = new Document();
-                $document->document_created_from_id = $documentCreatedFromParticipantId;
-                $document->document_type = 'internal';
-                $document->document_group = 'revenue';
-                $document->contact_id = $contact->id;
-                $document->project_id = $project->id;
-                $document->participation_project_id = $distributionKwh->participation_id;
-                $document->template_id = $documentTemplate->id;
-                $document->show_on_portal = $showOnPortal;
-
-                $filename = str_replace(' ', '', $this->translateToValidCharacterSet($project->code)) . '_'
-                    . str_replace(' ', '', $this->translateToValidCharacterSet($contact->full_name));
-
-
-                //max length name 25
-                $filename = substr($filename, 0, 25);
-
-                $document->filename = $filename
-                    . substr($document->getDocumentGroup()->name, 0, 1)
-                    . (Document::where('document_group', 'revenue')->count()
-                        + 1) . '_' . $time->format('Ymd') . '.pdf';
-
-                $document->save();
-
-                $filePath = (storage_path('app' . DIRECTORY_SEPARATOR
-                    . 'documents/' . $document->filename));
-                file_put_contents($filePath, $pdf);
-
-                if(\Config::get('app.ALFRESCO_COOP_USERNAME') != 'local') {
-                    $alfrescoHelper = new AlfrescoHelper(\Config::get('app.ALFRESCO_COOP_USERNAME'),
-                        \Config::get('app.ALFRESCO_COOP_PASSWORD'));
-
-                    $alfrescoResponse = $alfrescoHelper->createFile($filePath,
-                        $document->filename, $document->getDocumentGroup()->name);
-                    if($alfrescoResponse == null)
-                    {
-                        throw new \Exception('Fout bij maken rapport document in Alfresco.');
-                    }
-                    $document->alfresco_node_id = $alfrescoResponse['entry']['id'];
-                }else{
-                    $document->alfresco_node_id = null;
+                if ($documentTemplate->baseTemplate) {
+                    $html .= TemplateVariableHelper::replaceTemplateTagVariable($documentTemplate->baseTemplate->html_body,
+                        $documentTemplate->html_body, '', '');
+                } else {
+                    $html .= TemplateVariableHelper::replaceTemplateFreeTextVariables($documentTemplate->html_body,
+                        '', '');
                 }
 
-                $document->save();
-            }
-            catch (\Exception $e) {
-                Log::error('Fout bij maken rapport document voor ' . ($primaryEmailAddress ? $primaryEmailAddress->email : '**onbekend emailadres**') . ' (' . $contact->full_name . ')' );
-                Log::error($e->getMessage());
-                array_push($messages, 'Fout bij maken rapport document voor ' . ($primaryEmailAddress ? $primaryEmailAddress->email : '**onbekend emailadres**') . ' (' . $contact->full_name . ')' );
+                $html .= $documentTemplate->footer
+                    ? $documentTemplate->footer->html_body : '';
+
+                $subject = str_replace('{contactpersoon}', $contactInfo['contactPerson'], $subject);
+
+                $html = str_replace('{contactpersoon}', $contactInfo['contactPerson'], $html);
+
+                $revenueHtml = TemplateTableHelper::replaceTemplateTables($html, $contact);
+
+                $revenueHtml
+                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'contact', $contact);
+                $revenueHtml = TemplateVariableHelper::replaceTemplatePortalVariables($revenueHtml, 'portal');
+                $revenueHtml = TemplateVariableHelper::replaceTemplatePortalVariables($revenueHtml, 'contacten_portal');
+                $revenueHtml = TemplateVariableHelper::replaceTemplateCooperativeVariables($revenueHtml, 'cooperatie');
+
+                //wettelijk vertegenwoordiger
+                if (OccupationContact::where('contact_id', $contact->id)->where('occupation_id', 7)->exists()) {
+                    $wettelijkVertegenwoordiger = OccupationContact::where('contact_id', $contact->id)
+                        ->where('occupation_id', 7)->first()->primaryContact;
+                    $revenueHtml
+                        = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'wettelijk_vertegenwoordiger', $wettelijkVertegenwoordiger);
+                }
+                $revenueHtml
+                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'ik', $user);
+                $revenueHtml
+                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'administratie', $administration);
+                $revenueHtml
+                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'verdeling', $distributionKwh);
+                $revenueHtml
+                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'opbrengst', $revenuesKwh);
+                $revenueHtml
+                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'project', $project);
+                $revenueHtml
+                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'deelname', $distributionKwh->participation);
+                $revenueHtml
+                    = TemplateVariableHelper::replaceTemplateVariables($revenueHtml, 'mutaties', $distributionKwh->participation->mutations);
+
+                $revenueHtml
+                    = TemplateVariableHelper::stripRemainingVariableTags($revenueHtml);
+                $pdf = PDF::loadView('documents.generic', [
+                    'html' => $revenueHtml,
+                ])->output();
+
+                try {
+                    $time = Carbon::now();
+
+                    $documentCreatedFromParticipantId = DocumentCreatedFrom::where('code_ref', 'participant')->first()->id;
+
+                    $document = new Document();
+                    $document->document_created_from_id = $documentCreatedFromParticipantId;
+                    $document->document_type = 'internal';
+                    $document->document_group = 'revenue';
+                    $document->contact_id = $contact->id;
+                    $document->project_id = $project->id;
+                    $document->participation_project_id = $distributionKwh->participation_id;
+                    $document->template_id = $documentTemplate->id;
+                    $document->show_on_portal = $showOnPortal;
+
+                    $filename = str_replace(' ', '', $this->translateToValidCharacterSet($project->code)) . '_'
+                        . str_replace(' ', '', $this->translateToValidCharacterSet($contact->full_name));
+
+
+                    //max length name 25
+                    $filename = substr($filename, 0, 25);
+
+                    $document->filename = $filename
+                        . substr($document->getDocumentGroup()->name, 0, 1)
+                        . (Document::where('document_group', 'revenue')->count()
+                            + 1) . '_' . $time->format('Ymd') . '.pdf';
+
+                    $document->save();
+
+                    $filePath = (storage_path('app' . DIRECTORY_SEPARATOR
+                        . 'documents/' . $document->filename));
+                    file_put_contents($filePath, $pdf);
+
+                    if (\Config::get('app.ALFRESCO_COOP_USERNAME') != 'local') {
+                        $alfrescoHelper = new AlfrescoHelper(\Config::get('app.ALFRESCO_COOP_USERNAME'),
+                            \Config::get('app.ALFRESCO_COOP_PASSWORD'));
+
+                        $alfrescoResponse = $alfrescoHelper->createFile($filePath,
+                            $document->filename, $document->getDocumentGroup()->name);
+                        if ($alfrescoResponse == null) {
+                            throw new \Exception('Fout bij maken rapport document in Alfresco.');
+                        }
+                        $document->alfresco_node_id = $alfrescoResponse['entry']['id'];
+                    } else {
+                        $document->alfresco_node_id = null;
+                    }
+
+                    $document->save();
+                } catch (\Exception $e) {
+                    Log::error('Fout bij maken rapport document voor ' . ($primaryEmailAddress ? $primaryEmailAddress->email : '**onbekend emailadres**') . ' (' . $contact->full_name . ')');
+                    Log::error($e->getMessage());
+                    array_push($messages, 'Fout bij maken rapport document voor ' . ($primaryEmailAddress ? $primaryEmailAddress->email : '**onbekend emailadres**') . ' (' . $contact->full_name . ')');
+                }
             }
 
             //send email
@@ -805,7 +803,7 @@ class RevenuesKwhController extends ApiController
 
                     $email = Mail::fromMailbox($mailbox)
                         ->to($primaryEmailAddress->email);
-                    
+
                     if (!$subject) {
                         $subject = 'Participant rapportage Econobis';
                     }
@@ -869,7 +867,7 @@ class RevenuesKwhController extends ApiController
             }
 
             //delete file on server, still saved on alfresco.
-            if(\Config::get('app.ALFRESCO_COOP_USERNAME') != 'local') {
+            if($document && \Config::get('app.ALFRESCO_COOP_USERNAME') != 'local') {
                 Storage::disk('documents')->delete($document->filename);
             }
         }
